feat: add mobile sidebar section feedback
